+++ app/static/js/leaflet/_utils.js
@@ -1,5 +1,7 @@
 import * as L from "leaflet"
 
+const minBoundsSizePx = 20
+
 /**
  * Get a marker icon
  * @param {string} color Marker color/theme
@@ -34,4 +36,34 @@ export const getLatLngBoundsSize = (bounds) => {
     const sw = bounds.getSouthWest()
     const ne = bounds.getNorthEast()
     return (ne.lng - sw.lng) * (ne.lat - sw.lat)
+}
+
+/**
+ * Make bounds minimum size to make them easier to click
+ * @param {L.Map} map Leaflet map
+ * @param {number[]} bounds The bounds in [minLon, minLat, maxLon, maxLat]
+ * @returns {number[]} The new bounds
+ */
+export const makeBoundsMinimumSize = (map, bounds) => {
+    const [minLon, minLat, maxLon, maxLat] = bounds
+    const mapBottomLeft = map.project(L.latLng(minLat, minLon))
+    const mapTopRight = map.project(L.latLng(maxLat, maxLon))
+    const width = mapTopRight.x - mapBottomLeft.x
+    const height = mapBottomLeft.y - mapTopRight.y
+
+    if (width < minBoundsSizePx) {
+        const diff = minBoundsSizePx - width
+        mapBottomLeft.x -= diff / 2
+        mapTopRight.x += diff / 2
+    }
+
+    if (height < minBoundsSizePx) {
+        const diff = minBoundsSizePx - height
+        mapBottomLeft.y += diff / 2
+        mapTopRight.y -= diff / 2
+    }
+
+    const latLngBottomLeft = map.unproject(mapBottomLeft)
+    const latLngTopRight = map.unproject(mapTopRight)
+    return [latLngBottomLeft.lng, latLngBottomLeft.lat, latLngTopRight.lng, latLngTopRight.lat]
 }